pkg/generator: renaming Apis -> API
