+++ pkg/generator/gen_api_doc.go
@@ -1,0 +1,28 @@
+package generator
+
+import (
+	"io"
+	"text/template"
+)
+
+// Doc contains all the customized data needed to generate apis/<version>/doc.go for a new operator
+// when pairing with apisDocTmpl template.
+type Doc struct {
+	GroupName string
+	Version   string
+}
+
+// renderAPIDocFile generates the apis/<version>/doc.go file.
+func renderAPIDocFile(w io.Writer, groupName, version string) error {
+	t := template.New("apis/<version>/doc.go")
+	t, err := t.Parse(apiDocTmpl)
+	if err != nil {
+		return err
+	}
+
+	d := Doc{
+		GroupName: groupName,
+		Version:   version,
+	}
+	return t.Execute(w, d)
+}
